core: don't module_get() for internal buffer/trigger
If a driver module also create an internal trigger or buffer the ZIO
core increment the use counter of the module. Then the user cannot remove the
module because of the use counter. He must change the trigger or buffer with
another kind of trigger or buffer, and then he can remove the module.

This patch resolve this issue. Now the ZIO core don't increment the use counter
when a driver is using an internal trigger or buffer

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>
Acked-by: Alessandro Rubini <rubini@gnudd.com>