new feature: a "sniff" device receiving all control structures
/dev/zio-sniff.ctrl is a misc device, selected by CONFIG_ZIO_SNIFF_DEV
which currently is on by default.  Every time a block is released,
a copy of the control is delivered to every reader of the device.

This is meant to be useful for monitoring, and to retrieve the
perfomance figures that will be added by later commits.  We need
a sniff device to monitor output channels, because even the application
driving I/O cannot recover the control structures that leave the
ZIO pipeline.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>