bugfix: sysfs: change interaction of configuration and data I/O
When a process is writing and another process is changing the
configuration, it may happen that the trigger-related store calling
abort_disable() zeroes the active_block that is already being pushed
by the other process.

This fixes the interaction by keeping the cset spinlock during the whole
configuration of a sysfs attribute, thus preventing a write to sneak in.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Federico Vaga <federico.vaga@gmail.com>